[Substratum] OMS: Dismantle FLAG_OVERLAY_TRUSTED check in isOverlayPackage
This was added with commit 7a722dcae9ac but this appears to be an
incomplete attempt at doing what it should, because FLAG_OVERLAY_TRUSTED
is never assigned, only checked. While we wait for Google to realise this
and push a fuller commit for us to mitigate against, this bandaid does the
trick.

[harsh@prjkt.io: Add commit log]
Change-Id: Ic8668f4d2769b9b5daeea56cd6562674258a10b1
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>